Add integration test between provider and Mongo
We want to test the provider passes the correct parameters to the client
that will interact with AWS. In this case the client is for MongoDB,
but this provider could be extended to include other AWS services.

Options for testing this included:

* making `MongoDBService` conform to an interface and mock it
* not mocking the service and only testing the final integration (test
  AWS API gets the right parameters sent to it).

The first option was excluded as we cannot guarantee the behaviour for
all AWS service clients will be the same. For example, they may not use
CloudFormation and the orchestration may be very different. The
convenience of being able to mock the service didn't seem worth forcing
the interface upon future services.

I went with the latter option because it provides similar coverage and
gives confidence that a service plan configuration in the catalog
produces the correct CloudFormation stack.