Implement `LastOperation` in the provider
The operation type (for example, provision) and service (`mongodb`) have
been strictly checked here as we don't ever expect to receive any other
values. In theory, we could not enforce the operation type and just
check the state of the stack has some form of final status.  The problem
there would be the final state (for example, `ROLLBACK_COMPLETE`) may
not be the intended one (`CREATE_COMPLETE`).