Remove Terraform from integration tests
It was quite laborious to have to run Terraform manually and collect
outputs to set as environment variables for the test run. Here we
provide the same VPC setup, but with the addition of creating an SSH key
pair. The running of Terraform could have been moved into the
integration test, but that would mean a dependency on a third-party
tool. The infrastructure is simple enough to be managed with the SDK.

We have a configuration of private subnets for the MongoDB stack because
that is the expected configuration for the CloudFormation template.