Use Hex encoding for passwords, not base64
Base64 pads with the equals sign, which is not an allowed character in
the stack template for MongoDB. Currently the stack template uses an
embedded template maintained by AWS in S3, so we cannot edit it to allow
the use of the equals sign. We use Hex because it is URL safe and
contains only allowed characters.