Add AWS functions for updating MongoDB stacks
This provides a generic mechanism for updating a stack based on any
stack template parameter values. At this stage only updating the node
instance type is supported, but we allow any form of update because we
will defer responsibility to the service broker to decide what
parameters can be updated.